Move NewsImageDataLoaderPresentationAdapter to separate file
